+++ packages/core/src/main/java/dev/synapse/tools/ToolExecutionTimeoutException.java
@@ -1,0 +1,9 @@
+package dev.synapse.tools;
+
+import dev.synapse.core.infrastructure.exception.ApiException;
+
+public class ToolExecutionTimeoutException extends ApiException {
+    public ToolExecutionTimeoutException(String message) {
+        super("TOOL_EXECUTION_TIMEOUT", message, 408);
+    }
+}